requirements.txt
easy_vqa==1.0
gdown==3.9.0
matplotlib==3.4.3
numpy==1.20.3
pandas==1.3.4
Pillow==9.2.0
prettytable==3.3.0
pycocotools==2.0.4
pytorch_warmup==0.1.0
scikit_image==0.18.3
seaborn==0.11.2
requests==2.22.0
torch==1.12.1
torchvision==0.13.1
transformers==4.21.1
wandb
opencv-python==4.6.0.66 
